# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
...
  /api/v1/manage/clubs/{clubId}/teams/{teamId}/seasons/{seasonId}/squad/{playerId}:
    put:
      tags:
      - team-squad-controller
      operationId: update_3
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamSquadMemberJerseyNumberRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamSquadMemberDto"
...
  /api/v1/manage/clubs/{clubId}/teams/{teamId}/seasons/{seasonId}/squad/{playerId}/remove:
    post:
      tags:
      - team-squad-controller
      operationId: remove
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/manage/clubs/{clubId}/teams/{teamId}/seasons/{seasonId}/squad/{playerId}/add:
    post:
      tags:
      - team-squad-controller
      operationId: add
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamSquadMemberDto"
...
  /api/v1/manage/clubs/{clubId}/teams/{teamId}/seasons/{seasonId}/squad:
    get:
      tags:
      - team-squad-controller
      operationId: list_15
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: teamId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamSquadMemberDto"
...
components:
  schemas:
...
    UpdateTeamSquadMemberJerseyNumberRequest:
      type: object
      properties:
        jerseyNumber:
          type: integer
          format: int32
    TeamSquadMemberDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        playerProfileId:
          type: string
          format: uuid
        personId:
          type: string
          format: uuid
        clubId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        photoUrl:
          type: string
        clubMembershipNumber:
          type: string
        medicalAidProvider:
          type: string
        medicalAidMemberNumber:
          type: string
        phone:
          type: string
        email:
          type: string
        altContactName:
          type: string
        altContactPhone:
          type: string
        battingStance:
          type: string
          enum:
          - RIGHT_HANDED
          - LEFT_HANDED
        bowlingArm:
          type: string
          enum:
          - RIGHT_ARM
          - LEFT_ARM
        bowlingType:
          type: string
          enum:
          - FAST
          - FAST_MEDIUM
          - MEDIUM_FAST
          - MEDIUM
          - OFF_BREAK
          - LEG_BREAK
          - ORTHODOX_SPIN
          - WRIST_SPIN
          - GOOGLY
        isWicketKeeper:
          type: boolean
        active:
          type: boolean
        sectionIds:
          type: array
          items:
            type: string
            format: uuid
        jerseyNumber:
          type: integer
          format: int32
        squadJerseyNumber:
          type: integer
          format: int32
...